requirements.txt
fastapi
uvicorn
websockets
numpy
pillow
torch
opencv-python
facenet-pytorch
python-multipart
